+++ website/docusaurus.config.ts
@@ -24,6 +24,7 @@ const config: Config = {
   // If you aren't using GitHub pages, you don't need these.
   organizationName: 'usetheo',
   projectName: 'theo-openvoice',
+  trailingSlash: false,
 
   onBrokenLinks: 'throw',